renaming contexts.py to context_handler.py. first pass
